feat(db): migration framework
This introduces a framework for tracking a database's version and
migrating from one version to the next when the database schema needs
to be modified. The migrations consist of an array of methods that each
are responsible for upgrading from a particular version, and they are
run in sequence when we detect that the current database version is
lower than the latest version.